requirements.txt
pynput==1.7.6
watchdog==4.0.0
Pillow==10.2.0
google-genai==0.3.0
python-dotenv==1.0.1